Fix 1.3 incompatibility failures introduced during merge
