--- src/main/kotlin/com/fasterxml/jackson/module/kotlin/ConstructorValueCreator.kt
@@ -1,12 +1,0 @@
-package com.fasterxml.jackson.module.kotlin
-
-import kotlin.reflect.KFunction
-import kotlin.reflect.jvm.isAccessible
-
-internal class ConstructorValueCreator<T>(override val callable: KFunction<T>) : ValueCreator<T>() {
-    override val accessible: Boolean = callable.isAccessible
-
-    init {
-        callable.isAccessible = true
-    }
-}
