+++ src/testKotlin-1.5/kotlin/com/fasterxml/jackson/module/kotlin/test/github/GitHub524.kt
